Refactor Result and Results'2 classes for generic handling
The class Result is now handling T type instead of TSuccess. This modification allows to handle both success and failure states in a more generic way. Also, the IsSuccess property is now initialized true by default.

In the Results'2 class, the IsError property is placed under the IsSuccess property for readability, and some changes in the order of member variables were made to enhance code readability and structure.