Update project and package references for Debug/Release configurations
The project reference to 'OneOf.Reduced' was updated in 'StateResults.csproj'. For Debug configuration, it remains as a project reference and for Release configuration, it has been changed to a package reference 'BMTLab.OneOf.Reduced'. Simultaneously, in 'Directory.Packages.props', the 'OneOf' package was replaced with 'BMTLab.OneOf.Reduced'. This is done to ensure that the Release configuration uses the package from the nuget feed, providing more consistent and stable releases.