Refactor `OneOf` from a readonly record struct to a class
The `OneOf` type has been refactored from a readonly record struct to a class and now implements the `IEquatable` interface. This was changed to allow identity comparison operations (`==` and `!=` operators and `Equals` method). Also, added an optimized `GetHashCode` method implementation to correctly and more efficiently hash the state of the union. These changes are expected to increase the performance and ease of using `OneOf` instances in collections or for comparison operations.