Refactor the _index field to a property named Index
This commit redefines the _index field as a property named Index in the Results class, and uses this property throughout the class, updating all references to _index to now reference the Index property. The need for this refactoring arose to expose the index information outside the class without breaking encapsulation. It also creates a standard for accessing state data in the class increasing code maintainability.