Update project and package versions to release
Updated version number of the project in Directory.Build.props from 2.0 to 2.1 and removed the static package version. Added a condition to create a package version depending on the 'IsPreview' flag. This allows us to differentiate between preview and standard versions. Also updated BMTLab.OneOf.Reduced package version and JetBrains.Annotations version in Directory.Packages.props to newer versions. These changes are required to maintain up-to-date dependencies and to adjust the package versioning approach for better release management.