Update package and minor version
The BMTLab.OneOf.Reduced package has been updated from version 2.1.816 to 2.2.816 to include the latest features and updates. As a result of this package update, the project's minor version was incremented from 2.1 to 2.2.

These changes ensure our project stays current with external package updates, and the versioning accurately represents the project's evolution.